Split Core/Database into 2 files
